Split core lib in multiple files
