+++ lib/error-correction-level.js
@@ -1,0 +1,6 @@
+module.exports = {
+  L : 1,
+  M : 0,
+  Q : 3,
+  H : 2
+};
